Remove metricsRepository from ReportService
